integrations/servicenow2: move snow package to an internal directory
