+++ integrations/servicenow2/cmd/client.go
@@ -28,7 +28,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/itrs-group/cordial/integrations/servicenow2/snow"
+	"github.com/itrs-group/cordial/integrations/servicenow2/internal/snow"
 	"github.com/itrs-group/cordial/pkg/config"
 	"github.com/itrs-group/cordial/pkg/rest"
