+++ integrations/servicenow2/internal/snow/endpoints.go
